fix rpcclient getnetworkhashps JSON marshall error after handleGetNetworkHashPS() was changed from int64 to float64 type (#113)

+++ rpcserverhelp.go
@@ -706,7 +706,7 @@ var rpcResultTypes = map[string][]interface{}{
 	"getmempoolinfo":        {(*btcjson.GetMempoolInfoResult)(nil)},
 	"getmininginfo":         {(*btcjson.GetMiningInfoResult)(nil)},
 	"getnettotals":          {(*btcjson.GetNetTotalsResult)(nil)},
-	"getnetworkhashps":      {(*int64)(nil)},
+	"getnetworkhashps":      {(*float64)(nil)},
 	"getpeerinfo":           {(*[]btcjson.GetPeerInfoResult)(nil)},
 	"getrawmempool":         {(*[]string)(nil), (*btcjson.GetRawMempoolVerboseResult)(nil)},
 	"getrawtransaction":     {(*string)(nil), (*btcjson.TxRawResult)(nil)},
